fix: fixed some OTP handling functions
