+++ gateway/internal/application/usecase/otp_usecase.go
@@ -3,6 +3,6 @@ package usecase
 import "context"
 
 type OTPService interface {
-	GenerateOTP(ctx context.Context, phoneNumber string) error
+	GenerateAndStoreOTP(ctx context.Context, phoneNumber string) (plainOTP string, err error)
 	VerifyOTP(ctx context.Context, phoneNumber string, otp string) error
 }
