--- gateway/internal/application/usecase/sms_usecase.go
@@ -1,7 +1,0 @@
-package usecase
-
-import "context"
-
-type SMSService interface {
-	Send(ctx context.Context, phoneNumber string, message string) error
-}
