+++ gateway/internal/domain/communication/sms.go
@@ -1,28 +1,7 @@
 package communication
 
-import (
-	"context"
-	"log"
+import "context"
 
-	"github.com/Barbod-Biometrics/Backend/gateway/internal/application/usecase"
-)
-
-type smsService struct {
-	// client
-	// fromnumber string
-}
-
-func NewSMSService( /* config */ ) usecase.SMSService {
-	return &smsService{
-		// clident: client,
-		// fromnumber: fromnumber,
-	}
-
-}
-
-func (s *smsService) Send(ctx context.Context, phoneNumber string, message string) error {
-	// we'll send the OTP from here and save it in the redis
-	log.Printf("SMS to %s: %s\n", phoneNumber, message)
-
-	return nil
+type SMSService interface {
+	Send(ctx context.Context, phoneNumber string, message string) error
 }
